Keep only stack-neutral types in the common modules
A blocking-only port sat in a common module that the reactor module also
depends on, so the reactor stack transitively imported an interface it can
never call, and the blocking half got the unprefixed name by accident of
placement rather than by design.

TransactionExecutor and SynchronousEventDispatcher move from
application/service/common to application/service/blocking. ExecuteFilter
and dcb/TagGenerator stay, being genuinely stack-neutral.

SnapshotStore and InMemorySnapshotStore move from dsl/snapshot-dsl/common
to dsl/snapshot-dsl/blocking. SnapshotSupport splits along the same line:
the three store-taking methods (maybeSave and both maybeSaveBestEffort
overloads) move to the blocking module's internal package, while the pure
resolveBase, isRedelivery and requireInt stay in common/internal because
ReactiveSnapshotSupport reuses them.

The two SnapshotPolicies copies were identical apart from the package line
and depend on nothing stack-specific, so they collapse into one class in
dsl/snapshot-dsl/common.

This mirrors how application/service already places ApplicationService in
blocking and reactor with no prefix. spring-boot-starter-mongodb gains an
explicit application-service-blocking dependency, having previously picked
those types up transitively, and the two transaction adapters now depend on
application-service-blocking instead of common. That pulls
eventstore-api-blocking in transitively, which is accepted: there is no
cycle, and a transaction adapter is never used without the blocking
application service.

All of this is unreleased, so no migration recipe or upgrade-guide entry.
