+++ application/service/blocking/src/main/java/org/occurrent/application/service/blocking/generic/GenericApplicationService.java
@@ -22,8 +22,8 @@ import org.jspecify.annotations.Nullable;
 import org.occurrent.application.converter.CloudEventConverter;
 import org.occurrent.application.service.blocking.ApplicationService;
 import org.occurrent.application.service.ExecuteFilter;
-import org.occurrent.application.service.SynchronousEventDispatcher;
-import org.occurrent.application.service.TransactionExecutor;
+import org.occurrent.application.service.blocking.SynchronousEventDispatcher;
+import org.occurrent.application.service.blocking.TransactionExecutor;
 import org.occurrent.application.service.blocking.ExecuteOptions;
 import org.occurrent.eventstore.api.StreamReadFilter;
 import org.occurrent.eventstore.api.WriteConditionNotFulfilledException;
