+++ application/service/transaction-mongodb-native/pom.xml
@@ -31,7 +31,7 @@
     <dependencies>
         <dependency>
             <groupId>org.occurrent</groupId>
-            <artifactId>occurrent-application-service-common</artifactId>
+            <artifactId>occurrent-application-service-blocking</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
@@ -50,12 +50,6 @@
         </dependency>
 
         <!-- Test -->
-        <dependency>
-            <groupId>org.occurrent</groupId>
-            <artifactId>occurrent-application-service-blocking</artifactId>
-            <version>${project.version}</version>
-            <scope>test</scope>
-        </dependency>
         <dependency>
             <groupId>org.occurrent</groupId>
             <artifactId>occurrent-subscription-synchronous-blocking</artifactId>
