+++ dsl/snapshot-dsl/blocking/src/main/java/org/occurrent/dsl/snapshot/blocking/SnapshotStore.java
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package org.occurrent.dsl.snapshot;
+package org.occurrent.dsl.snapshot.blocking;
 
 import org.jspecify.annotations.Nullable;
+import org.occurrent.dsl.snapshot.Snapshot;
 
 import java.util.Optional;
