+++ dsl/snapshot-dsl/blocking/src/test/java/org/occurrent/dsl/snapshot/blocking/SnapshotDeciderApplicationServiceTest.java
@@ -33,8 +33,8 @@ import org.occurrent.domain.NameWasChanged;
 import org.occurrent.dsl.decider.Decider;
 import org.occurrent.dsl.snapshot.Snapshot;
 import org.occurrent.dsl.snapshot.SnapshotOptions;
+import org.occurrent.dsl.snapshot.SnapshotPolicies;
 import org.occurrent.dsl.snapshot.SnapshotPolicy;
-import org.occurrent.dsl.snapshot.SnapshotStore;
 import org.occurrent.eventstore.api.WriteResult;
 import org.occurrent.eventstore.inmemory.InMemoryEventStore;
