+++ dsl/snapshot-dsl/common/src/test/java/org/occurrent/dsl/snapshot/SnapshotSupportTest.java
@@ -17,11 +17,8 @@
 package org.occurrent.dsl.snapshot;
 
 import org.junit.jupiter.api.Test;
-import org.occurrent.dsl.snapshot.LedgerFixture.Deposited;
-import org.occurrent.dsl.snapshot.LedgerFixture.LedgerEvent;
 import org.occurrent.dsl.snapshot.internal.SnapshotSupport;
 
-import java.util.List;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -54,28 +51,6 @@ class SnapshotSupportTest {
 
         assertThat(base.state()).isEqualTo(5);
         assertThat(base.version()).isEqualTo(0);
-    }
-
-    @Test
-    void maybe_save_writes_a_tagged_snapshot_when_the_policy_fires() {
-        SnapshotStore<Integer> store = SnapshotStore.inMemory();
-        SnapshotDecision<Integer, LedgerEvent> decision = new SnapshotDecision<>(42, List.of(new Deposited(1)), 7, 1);
-
-        boolean saved = SnapshotSupport.maybeSave(store, "acc:1", 3, SnapshotPolicy.always(), decision);
-
-        assertThat(saved).isTrue();
-        assertThat(store.findLatest("acc:1")).contains(new Snapshot<>(42, 7, 3));
-    }
-
-    @Test
-    void maybe_save_writes_nothing_when_the_policy_does_not_fire() {
-        SnapshotStore<Integer> store = SnapshotStore.inMemory();
-        SnapshotDecision<Integer, LedgerEvent> decision = new SnapshotDecision<>(42, List.of(new Deposited(1)), 7, 1);
-
-        boolean saved = SnapshotSupport.maybeSave(store, "acc:1", 3, SnapshotPolicy.never(), decision);
-
-        assertThat(saved).isFalse();
-        assertThat(store.findLatest("acc:1")).isEmpty();
     }
 
     @Test
@@ -136,18 +111,5 @@ class SnapshotSupportTest {
 
         // The head is at 7, the snapshot is at 7, and version 5 was already folded: a genuine redelivery to skip.
         assertThat(SnapshotSupport.isRedelivery(loaded, 1, 5, 7)).isTrue();
-    }
-
-    @Test
-    void best_effort_save_swallows_a_negative_events_since_snapshot_rather_than_throwing() {
-        SnapshotStore<Integer> store = SnapshotStore.inMemory();
-
-        // The supplier assembles the decision with a negative eventsSinceSnapshot, which requireInt rejects. Building it
-        // inside the best-effort boundary means the throw is swallowed (the command already committed), not propagated.
-        boolean saved = SnapshotSupport.<Integer, LedgerEvent>maybeSaveBestEffort(store, "acc:1", 3, SnapshotPolicy.always(),
-                () -> new SnapshotDecision<>(42, List.of(new Deposited(1)), 7, SnapshotSupport.requireInt(-1, "the number of events since the snapshot")));
-
-        assertThat(saved).isFalse();
-        assertThat(store.findLatest("acc:1")).isEmpty();
     }
 }
