+++ dsl/snapshot-dsl/reactor/src/test/java/org/occurrent/dsl/snapshot/reactor/ReactiveSnapshotDeciderApplicationServiceTest.java
@@ -37,6 +37,7 @@ import org.occurrent.domain.NameWasChanged;
 import org.occurrent.dsl.decider.Decider;
 import org.occurrent.dsl.snapshot.Snapshot;
 import org.occurrent.dsl.snapshot.SnapshotOptions;
+import org.occurrent.dsl.snapshot.SnapshotPolicies;
 import org.occurrent.dsl.snapshot.SnapshotPolicy;
 import org.occurrent.eventstore.mongodb.spring.reactor.EventStoreConfig;
 import org.occurrent.eventstore.mongodb.spring.reactor.ReactorMongoEventStore;
