+++ example/snapshot/closing-the-books/src/main/kotlin/org/occurrent/example/snapshot/closingthebooks/ClosingTheBooksDemo.kt
@@ -21,10 +21,10 @@ import org.occurrent.application.converter.typemapper.ReflectionCloudEventTypeMa
 import org.occurrent.application.service.blocking.generic.GenericApplicationService
 import org.occurrent.dsl.snapshot.SnapshotOptions
 import org.occurrent.dsl.snapshot.SnapshotPolicy
-import org.occurrent.dsl.snapshot.SnapshotStore
+import org.occurrent.dsl.snapshot.blocking.SnapshotStore
 import org.occurrent.dsl.snapshot.blocking.SnapshotDecider
 import org.occurrent.dsl.snapshot.blocking.SnapshotDeciderApplicationService
-import org.occurrent.dsl.snapshot.blocking.SnapshotPolicies
+import org.occurrent.dsl.snapshot.SnapshotPolicies
 import org.occurrent.eventstore.inmemory.InMemoryEventStore
 import tools.jackson.module.kotlin.jacksonObjectMapper
 import java.net.URI
