+++ framework/spring-boot-starter-mongodb/src/main/java/org/occurrent/springboot/mongo/blocking/OccurrentMongoAutoConfiguration.java
@@ -24,8 +24,8 @@ import org.jspecify.annotations.NonNull;
 import org.occurrent.application.converter.CloudEventConverter;
 import org.occurrent.application.converter.typemapper.CloudEventTypeMapper;
 import org.occurrent.application.converter.typemapper.ReflectionCloudEventTypeMapper;
-import org.occurrent.application.service.SynchronousEventDispatcher;
-import org.occurrent.application.service.TransactionExecutor;
+import org.occurrent.application.service.blocking.SynchronousEventDispatcher;
+import org.occurrent.application.service.blocking.TransactionExecutor;
 import org.occurrent.application.service.blocking.ApplicationService;
 import org.occurrent.application.service.blocking.dcb.DcbApplicationService;
 import org.occurrent.application.service.blocking.dcb.GenericDcbApplicationService;
